fix: resolve type comparison error in txmempool.cpp
Add explicit cast to resolve sign comparison warning between int64_t
GetSizeWithAncestors() and uint64_t ancestor_size_limit parameter.

This addresses the compilation error:
comparison of integer expressions of different signedness: 'int64_t' and 'uint64_t'

Resolves CI build failures across multiple platforms.

+++ src/txmempool.cpp
@@ -128,7 +128,7 @@ void CTxMemPool::UpdateForDescendants(txiter updateIt, cacheMap& cachedDescendan
             // Don't directly remove the transaction here -- doing so would
             // invalidate iterators in cachedDescendants. Mark it for removal
             // by inserting into descendants_to_remove.
-            if (descendant.GetCountWithAncestors() > ancestor_count_limit || descendant.GetSizeWithAncestors() > ancestor_size_limit) {
+            if (descendant.GetCountWithAncestors() > ancestor_count_limit || descendant.GetSizeWithAncestors() > int64_t(ancestor_size_limit)) {
                 descendants_to_remove.insert(descendant.GetTx().GetHash());
             }
         }
